Revert "refactor of psychological test" Realised, it worked fine before:) This reverts commit 551781c061c337f014809ad93da746afdee7c57a.
