+++ src/main/java/ch/uzh/ifi/imrg/patientapp/rest/mapper/PsychologicalTestMapper.java
@@ -1,13 +1,14 @@
 package ch.uzh.ifi.imrg.patientapp.rest.mapper;
 
-import ch.uzh.ifi.imrg.patientapp.entity.PsychologicalTest.PsychologicalTest;
-import PsychologicalTestQuestions;
+import ch.uzh.ifi.imrg.patientapp.entity.PsychologicalTest;
+import ch.uzh.ifi.imrg.patientapp.entity.PsychologicalTestQuestions;
 import ch.uzh.ifi.imrg.patientapp.rest.dto.input.PsychologicalTestInputDTO;
 import ch.uzh.ifi.imrg.patientapp.rest.dto.input.PsychologicalTestQuestionInputDTO;
 import ch.uzh.ifi.imrg.patientapp.rest.dto.output.PsychologicalTestOutputDTO;
 import ch.uzh.ifi.imrg.patientapp.rest.dto.output.PsychologicalTestQuestionOutputDTO;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
+import org.mapstruct.Mappings;
 import org.mapstruct.factory.Mappers;
 
 import java.util.List;
